Add debian build
This adds a debian build, producing the following packages:

rpcd
rpcd-dev
rpcd-mod-file
rpcd-mod-iwinfo
rpcd-mod-rpcsys

rpc-mod-ucode is not yet available as ucode is not yet packaged. This relies
on the recent debian/ additions to libubox and ubus, as well as the following
pull requests adding additional debian packaging:

https://github.com/openwrt/iwinfo/pull/28

This also adds a systemd service file. Note that local ubus access provides
no permissions structure other than permissions on the socket itself. Be
sure to be aware of the implications.