file: bump sz_size to 64 bits
Some systems reveal /proc/kcore which for 64 bit systems has a size of

140,737,471,586,304 bytes.

This is a virtual mapping of the maximum possible RAM in a system. But if a 32
bit value tries to hold this, it is displayed as the erroneous value

-16,769,024 bytes

Bump the size to 64 bits to hold the correct value (and correct values for large
files).

Before this patch:

ubus call file list '{"path":"/proc"}'
...
{
	"name": "kcore",
	"type": "file",
	"size": -16769024,
	"mode": 33024,
	"atime": 1764716086,
	"mtime": 1764716086,
	"ctime": 1764716086,
	"inode": -268435271,
	"uid": 0,
	"gid": 0,
},
...

After this patch:

ubus call file list '{"path":"/proc"}'
...
{
	"name": "kcore",
	"type": "file",
	"size": 140737471586304,
	"mode": 33024,
	"atime": 1764716086,
	"mtime": 1764716086,
	"ctime": 1764716086,
	"inode": -268435271,
	"uid": 0,
	"gid": 0,
},
...

Tested on: OpenWrt SNAPSHOT r32139-1f879b8839 / 6.12.59
Signed-off-by: Paul Donald <newtwen+github@gmail.com>
Link: https://github.com/openwrt/rpcd/pull/22
Signed-off-by: Robert Marko <robimarko@gmail.com>